+++ scopes/compositions/compositions/index.ts
@@ -2,7 +2,7 @@ import { CompositionsAspect } from './compositions.aspect';
 
 export { CompositionsAspect };
 export type { CompositionsMain } from './compositions.main.runtime';
-export type { CompositionsUI } from './compositions.ui.runtime';
+export type { CompositionsUI, CompositionsMenuSlot } from './compositions.ui.runtime';
 export { Composition, CompositionProps } from './composition';
 export type { CompositionsPreview } from './compositions.preview.runtime';
 export { CompositionsOverview } from './ui/compositions-overview';
