+++ scopes/compositions/ui/hooks/use-composition/index.ts
@@ -1,0 +1,2 @@
+export { CompositionContextProvider, CompositionContextProviderProps, usePreview } from './use-composition';
+export { usePreviewQueryParams } from './use-preview-query-params';
