--- scopes/compositions/ui/use-composition/index.ts
@@ -1,6 +1,0 @@
-export {
-  CompositionContextProvider,
-  CompositionContextProviderProps,
-  usePreview,
-  usePreviewQueryParams,
-} from './use-composition';
