+++ scopes/react/bit-react-transformer/bit-id-from-pkg-json.ts
